refactor: split scripts into modules
