+++ index.html
@@ -47,8 +47,8 @@
           
   
     </div>
-    <!-- Seu link de script para o seu arquivo JavaScript personalizado -->
-    <script src="sinusite.js"></script>
+    <!-- Módulo principal da aplicação -->
+    <script type="module" src="ui.js"></script>
